Rebase with gradle build system and made TippieLogger more accessable
